% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quan_MMEJ.R
\name{Detect.MMEJ}
\alias{Detect.MMEJ}
\title{Detect MMEJ sequence from indel results}
\usage{
Detect.MMEJ(x, data = NULL, pb = NULL)
}
\arguments{
\item{x}{Number of rows}

\item{data}{data.frame of Alleles frequency table}

\item{pb}{ProgressBar object, setting by 'utils::txtProgressBar'}
}
\description{
Detect MMEJ sequence from indel results
}
