% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ampseq}
\alias{get_ampseq}
\title{get amplicon sequence from json file or log file of run parameters}
\usage{
get_ampseq(path)
}
\arguments{
\item{path}{path to the sample directory which generated by CRISPResso2}
}
\value{
The sequence of \code{amplicon}, that is reference sequence.
}
\description{
get amplicon sequence from json file or log file of run parameters
}
