% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_MCV}
\alias{plot_MCV}
\title{Plotting the modification count vectors}
\usage{
plot_MCV(path, start_pos, end_pos, show_gRNA = T)
}
\arguments{
\item{path}{path to the sample directory which generated by CRISPResso2}

\item{start_pos}{start position of the highlited mutation window}

\item{end_pos}{end position of the highlited mutation window}
}
\description{
Plotting the modification count vectors
}
