% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quan_MMEJ.R
\name{Quant_MMEJ}
\alias{Quant_MMEJ}
\title{Quant MMEJ sequence and length}
\usage{
Quant_MMEJ(path, nCores = 1, proportion = "mmej")
}
\arguments{
\item{path}{path to the sample directory which generated by CRISPResso2}

\item{nCores}{cores used to process data}

\item{proportion}{how to calculate the proportion of reads, select from c("mmej", "indel")}
}
\description{
Quant MMEJ sequence and length
}
